Fix hash calculation for catalog item of link
`fime.symLinkTarget()` of some *.lnk files returns empty string on
Windows 10.  For example, Window Switcher.lnk, Shows Desktop.lnk
and Sticky Notes.lnk.
Hash is 0 and items are not added to catalog properly since hash
is used for item equality check.
As a workaround we calculate hash using `fullPath` if the link
target is empty.